B2B Buyer Portal integration - cart synchronization (#2242)
* feat: Add support for Buyer Portal in Catalyst

* refactor: remove unnecessary useState/useEffects + simplify useB2B* logic

* refactor: simplify use of env. variables

* refactor: Simplify integration points with product-details-form

Use an onSubmit listener and move most of the logic outside the
product-details-form component.

The ideal solution here is to patch the form action at the root of the
component tree, but that's proven complicated at this point. Will try to
amend it in the future.

* fix: cart sync

* fix: deprecating sync-cart

* feat: Add support for Buyer Portal in Catalyst

* refactor: remove unnecessary useState/useEffects + simplify useB2B* logic

* refactor: simplify use of env. variables

* refactor: Simplify integration points with product-details-form

Use an onSubmit listener and move most of the logic outside the
product-details-form component.

The ideal solution here is to patch the form action at the root of the
component tree, but that's proven complicated at this point. Will try to
amend it in the future.

* fix: changes according to review

* fix: changes according to review

* fix: improving cartId prop type

---------

Co-authored-by: Ignacio Catalina <ignacio.catalina@bigcommerce.com>